+++ lib/event_handler.ex
@@ -38,9 +38,11 @@ defmodule GraphConn.EventHandler do
 
       @doc false
       @spec _request_cache_name() :: module()
-      def _request_cache_name(),
+      def _request_cache_name,
         do: Module.concat(__MODULE__, RequestCache)
 
+      @doc "Starts the event handler supervision tree."
+      @spec start_link(config :: nil | Keyword.t()) :: Supervisor.on_start()
       def start_link(config \\ nil) do
         Supervisor.start_link(__MODULE__, _put_config(config), name: __MODULE__)
       end
@@ -106,14 +108,17 @@ defmodule GraphConn.EventHandler do
       def handle_message(:"events-ws", msg, internal_state),
         do: handle_message_default(:"events-ws", msg, internal_state)
 
+      @doc false
       def handle_event_default(msg), do: handle_message_default(:"events-ws", msg, nil)
 
+      @doc false
       def handle_message_default(:"events-ws", msg, _) do
         Logger.warning(
           "[EventHandler] Received unexpected message from events-ws: #{inspect(msg)}"
         )
       end
 
+      @doc "Sends a `register` message to the events-ws upstream with `args` as the filter."
       def register(args) do
         GraphConn.execute(__MODULE__, :"events-ws", %GraphConn.Request{
           body: %{
@@ -123,7 +128,8 @@ defmodule GraphConn.EventHandler do
         })
       end
 
-      def subscribe() do
+      @doc "Sends a `subscribe` message to the events-ws upstream using the configured `scope_id`."
+      def subscribe do
         GraphConn.execute(__MODULE__, :"events-ws", %GraphConn.Request{
           body: %{
             type: "subscribe",
@@ -132,12 +138,15 @@ defmodule GraphConn.EventHandler do
         })
       end
 
+      @doc false
       def default_execution_timeout(_event),
         do: 60_000
 
+      @doc false
       def resend_response_timeout,
         do: 3_000
 
+      @doc false
       def handle_event(event) do
         Logger.warning("[EventHandler] Unhandled event message: #{event}")
         {:ok, :ignored}